End-to-End-Projects/Medical-Chatbot/requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
# langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
-e . 